+++ packages/plugins/type-distributor/src/headless/stores/dnd/dnd.store.svelte.ts
@@ -11,8 +11,8 @@ import {
 	generateCommitTitle,
 	commitEdits
 } from './drop-handler'
-import { assignedLNodesStore, bayStore } from '@/headless/stores'
-import { buildEditsForBayLNode } from '@/headless/ied'
+import { assignedLNodesStore } from '@/headless/stores'
+import { buildEditsForBayLNode } from '@/headless/scl'
 
 type DraggedItem = {
 	type: 'equipmentFunction' | 'functionTemplate' | 'lNode'