+++ packages/plugins/type-distributor/src/headless/stores/dnd/drop-handler.spec.ts
@@ -54,7 +54,7 @@ vi.mock('@/headless/utils/get-document-and-Editor', () => ({
 	getDocumentAndEditor: vi.fn()
 }))
 
-vi.mock('@/headless/ied', () => ({
+vi.mock('@/headless/scl', () => ({
 	createMultipleLNodesInAccessPoint: vi.fn()
 }))