+++ packages/plugins/type-distributor/src/ui/components/columns/s-ied/create-ied-ap-dialog/form-elements/multi-ap-back-button.svelte
@@ -1,0 +1,15 @@
+<script lang="ts">
+import { ArrowLeft } from '@lucide/svelte'
+import { Button } from '@oscd-plugins/core-ui-svelte'
+
+let {
+	onGoBackToIedSelection
+}: {
+	onGoBackToIedSelection: () => void
+} = $props()
+</script>
+
+<Button.Root variant="ghost" class="w-fit text-primary" onclick={onGoBackToIedSelection}>
+ <ArrowLeft></ArrowLeft>
+ Back
+</Button.Root>