+++ packages/plugins/type-distributor/src/ui/components/columns/s-ied/create-ied-ap-dialog/form-sections/index.ts
@@ -1,0 +1,3 @@
+export { default as IedSelectorSection } from './ied-selector-section.svelte'
+export { default as IedFormSection } from './ied-form-section.svelte'
+export { default as AccessPointFormSection } from './access-point-form-section.svelte'